+++ inc/Constants.php
@@ -17,13 +17,10 @@ use CrowdSecBouncer\Constants as LibConstants;
 class Constants extends LibConstants
 {
 
-    public const CROWDSEC_LOG_BASE_PATH = __DIR__ . '/../logs/';
-    public const CROWDSEC_LOG_PATH = __DIR__ . '/../logs/prod.log';
-    public const CROWDSEC_DEBUG_LOG_PATH = __DIR__ . '/../logs/debug.log';
-    public const CROWDSEC_CACHE_PATH = __DIR__ . '/../.cache';
-    public const CROWDSEC_CONFIG_PATH = __DIR__ . '/standalone-settings.php';
-    public const CROWDSEC_BOUNCER_USER_AGENT = 'WordPress CrowdSec Bouncer/v1.11.0';
-    public const CROWDSEC_BOUNCER_GEOLOCATION_DIR = __DIR__ . '/../geolocation';
-    public const CROWDSEC_BOUNCER_TLS_DIR = __DIR__ . '/../tls';
-
+    public const LOG_BASE_PATH = __DIR__ . '/../logs/';
+    public const CACHE_PATH = __DIR__ . '/../.cache';
+    public const CONFIG_PATH = __DIR__ . '/standalone-settings.php';
+    public const VERSION = 'v2.0.0';
+    public const GEOLOCATION_DIR = __DIR__ . '/../geolocation';
+    public const TLS_DIR = __DIR__ . '/../tls';
 }