+++ inc/options-config.php
@@ -1,7 +1,7 @@
 <?php
 require_once __DIR__ . '/Constants.php';
 
-function getCrowdSecOptionsConfig()
+function getCrowdSecOptionsConfig(): array
 {
     return [
         ['name' => 'crowdsec_api_url', 'default' => '', 'autoInit' => true],
@@ -23,8 +23,6 @@ function getCrowdSecOptionsConfig()
         ['name' => 'crowdsec_clean_ip_cache_duration', 'default' => Constants::CACHE_EXPIRATION_FOR_CLEAN_IP, 'autoInit' => true],
         ['name' => 'crowdsec_bad_ip_cache_duration', 'default' => Constants::CACHE_EXPIRATION_FOR_BAD_IP, 'autoInit' => true],
         ['name' => 'crowdsec_captcha_cache_duration', 'default' => Constants::CACHE_EXPIRATION_FOR_CAPTCHA,
-            'autoInit' => true],
-        ['name' => 'crowdsec_geolocation_cache_duration', 'default' => Constants::CACHE_EXPIRATION_FOR_GEO,
             'autoInit' => true],
         ['name' => 'crowdsec_fallback_remediation', 'default' => Constants::REMEDIATION_CAPTCHA, 'autoInit' => true],
         ['name' => 'crowdsec_hide_mentions', 'default' => '', 'autoInit' => true],
@@ -56,9 +54,21 @@ function getCrowdSecOptionsConfig()
         ['name' => 'crowdsec_forced_test_ip', 'default' => '', 'autoInit' => true],
         ['name' => 'crowdsec_forced_test_forwarded_ip', 'default' => '', 'autoInit' => true],
         ['name' => 'crowdsec_geolocation_enabled', 'default' => '', 'autoInit' => true],
-        ['name' => 'crowdsec_geolocation_save_result', 'default' => '', 'autoInit' => true],
+        ['name' => 'crowdsec_geolocation_cache_duration', 'default' => Constants::CACHE_EXPIRATION_FOR_GEO,
+            'autoInit' => true],
         ['name' => 'crowdsec_geolocation_type', 'default' => Constants::GEOLOCATION_TYPE_MAXMIND, 'autoInit' => true],
         ['name' => 'crowdsec_geolocation_maxmind_database_type', 'default' => Constants::MAXMIND_COUNTRY, 'autoInit' => true],
         ['name' => 'crowdsec_geolocation_maxmind_database_path', 'default' => '', 'autoInit' => true],
     ];
+}
+
+function getDatabaseConfigs(): array
+{
+    $crowdSecWpPluginOptions = getCrowdSecOptionsConfig();
+    $finalConfigs = [];
+    foreach ($crowdSecWpPluginOptions as $option) {
+        $finalConfigs[$option['name']] = get_option($option['name']);
+    }
+
+    return $finalConfigs;
 }